requirements.txt
antspyx
dcm2niix
dicom_csv
pydicom